requirements.txt
django
django-phonenumber-field phonenumbers
pillow
psycopg2-binary
qrcode